+++ run_test.sh
@@ -1,12 +1,12 @@
 #!/bin/bash 
 
-echo  "Flutterテスト実行中..." 
+echo  "Running Flutter tests..." 
 
 if [ " $1 " = "--coverage" ]; then
      flutter test --coverage 
-    echo  "除外されたファイルを削除..."
+    echo  "Deleting excluded files..."
      lcov -r coverage/lcov.info "lib/**.g.dart" -o coverage/lcov.info 
-    echo  "htmlファイルを生成しています..."
+    echo  "Generating html file..."
      genhtml coverage/lcov.info -o coverage/html 
 else
      flutter test